feat(universal-token): unified runtime input/output and improved encoding/decoding of storage and events

+++ crates/universal-token/src/types/derived_key.rs
@@ -2,7 +2,7 @@ use alloc::vec;
 use fluentbase_sdk::crypto::crypto_keccak256;
 use fluentbase_sdk::{B256, U256};
 
-pub type SlotType = u32;
+pub type SlotType = u8;
 
 const SLOT_BYTES: usize = size_of::<SlotType>();
